+++ terrareg/server.py
@@ -10,7 +10,7 @@ import hashlib
 from enum import Enum
 
 from flask import (
-    Flask, request, render_template,
+    Config, Flask, request, render_template,
     redirect, make_response, send_from_directory,
     session, g
 )
@@ -275,7 +275,7 @@ class Server(object):
         )
         self._api.add_resource(
             ApiTerraregGlobalUsageStats,
-            '/v1/terrareg/analaytics/global/usage_stats'
+            '/v1/terrareg/analytics/global/usage_stats'
         )
         self._api.add_resource(
             ApiTerraregModuleProviderAnalyticsTokenVersions,
@@ -334,6 +334,10 @@ class Server(object):
         self._api.add_resource(
             ApiModuleVersionCreateBitBucketHook,
             '/v1/terrareg/modules/<string:namespace>/<string:name>/<string:provider>/hooks/bitbucket'
+        )
+        self._api.add_resource(
+            ApiModuleVersionCreateGitHubHook,
+            '/v1/terrareg/modules/<string:namespace>/<string:name>/<string:provider>/hooks/github'
         )
         self._api.add_resource(
             ApiTerraregModuleVersionVariableTemplate,
@@ -1044,6 +1048,101 @@ class ApiModuleVersionCreateBitBucketHook(ErrorCatchingResource):
                 'tags': imported_versions
             }
 
+
+class ApiModuleVersionCreateGitHubHook(ErrorCatchingResource):
+    """Provide interface for GitHub hook to detect new and changed releases."""
+
+    def _post(self, namespace, name, provider):
+        """Create, update or delete new version based on GitHub release hooks."""
+        with Database.start_transaction():
+            namespace = Namespace(name=namespace)
+            module = Module(namespace=namespace, name=name)
+            # Get module provider and optionally create, if it doesn't exist
+            module_provider = ModuleProvider.get(
+                module=module,
+                name=provider,
+                create=terrareg.config.Config().AUTO_CREATE_MODULE_PROVIDER)
+
+            if not module_provider:
+                return self._get_404_response()
+
+            # Validate signature
+            if terrareg.config.Config().UPLOAD_API_KEYS:
+                # Get signature from request
+                request_signature = request.headers.get('X-Hub-Signature', '')
+                # Remove 'sha256=' from beginning of header
+                request_signature = re.sub(r'^sha256=', '', request_signature)
+                # Iterate through each of the keys and test
+                for test_key in terrareg.config.Config().UPLOAD_API_KEYS:
+                    # Generate
+                    valid_signature = hmac.new(bytes(test_key, 'utf8'), b'', hashlib.sha256)
+                    valid_signature.update(request.data)
+                    # If the signatures match, break from loop
+                    if hmac.compare_digest(valid_signature.hexdigest(), request_signature):
+                        break
+                # If a valid signature wasn't found with one of the configured keys,
+                # return 401
+                else:
+                    return self._get_401_response()
+
+            if not module_provider.get_git_clone_url():
+                return {'status': 'Error', 'message': 'Module provider is not configured with a repository'}, 400
+
+            github_data = request.json
+
+            if not ('release' in github_data and type(github_data['release']) == dict):
+                return {'status': 'Error', 'message': 'Received a non-release hook request'}, 400
+
+            release = github_data['release']
+    
+            # Obtain tag name
+            tag_ref = release.get('tag_name')
+            if not tag_ref:
+                return {'status': 'Error', 'message': 'tag_name not present in request'}, 400
+
+            # Attempt to match version against regex
+            version = module_provider.get_version_from_tag(tag_ref)
+
+            if not version:
+                return {'status': 'Error', 'message': 'Release tag does not match configured version regex'}, 400
+
+            # Create module version
+            module_version = ModuleVersion(module_provider=module_provider, version=version)
+
+            action = github_data.get('action')
+            if not action:
+                return {"status": "Error", "message": "No action present in request"}, 400
+
+            if action in ['deleted', 'unpublished']:
+                if not terrareg.config.Config().UPLOAD_API_KEYS:
+                    return {
+                        'status': 'Error',
+                        'message': 'Version deletion requires API key authentication',
+                        'tag': tag_ref
+                    }, 400
+                module_version.delete()
+
+                return {
+                    'status': 'Success'
+                }
+            else:
+                # Perform import from git
+                try:
+                    module_version.prepare_module()
+                    with GitModuleExtractor(module_version=module_version) as me:
+                        me.process_upload()
+
+                    return {
+                        'status': 'Success',
+                        'message': 'Imported provided tag',
+                        'tag': tag_ref
+                    }
+                except TerraregError as exc:
+                    return {
+                        'status': 'Error',
+                        'message': 'Tag failed to import',
+                        'tag': tag_ref
+                    }, 500
 
 class ApiModuleList(ErrorCatchingResource):
     def _get(self):